Ad i4hidden-text shortcode
Rational: The official hidden-text shortcode from rrze-elements allows
to hide parts of a website based on temporal annotations. However, it
might be desireable to hide those parts only externally. For instance,
it might be interesting to check formatting/timetables etc BEFORE
publication.

This shortcode is a verbatim copy of the official plugin, with the
modification that it allows logged in users (i.e. members of the chair)
to always view the hidden text.